requirements.txt
bootstrap4==0.1.0
click==8.1.3
Django==4.1.5
django-bootstrap-datepicker-plus==5.0.3
django-chartjs==2.3.0
djangorestframework==3.14.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
python-dotenv==0.21.0
python-telegram-bot==20.1
pytz==2022.7
requests==2.24.0

urllib3~=1.25.11